type consist

+++ analytical_engine/java/giraph-on-grape/src/main/java/com/alibaba/graphscope/samples/MessageBenchMark.java
@@ -13,7 +13,7 @@ import org.slf4j.LoggerFactory;
  * Only send msg.
  */
 public class MessageBenchMark extends
-    BasicComputation<LongWritable, DoubleWritable, DoubleWritable, LongWritable> {
+    BasicComputation<LongWritable, DoubleWritable, DoubleWritable, DoubleWritable> {
 
     private static Logger logger = LoggerFactory.getLogger(MessageBenchMark.class);
 
@@ -28,14 +28,14 @@ public class MessageBenchMark extends
      */
     @Override
     public void compute(Vertex<LongWritable, DoubleWritable, DoubleWritable> vertex,
-        Iterable<LongWritable> messages) throws IOException {
+        Iterable<DoubleWritable> messages) throws IOException {
         if (getSuperstep() >= MAX_SUPER_STEP){
             vertex.voteToHalt();
             return ;
         }
         if (getSuperstep() > 1){
             int msgCnt = 0;
-            for (LongWritable message : messages){
+            for (DoubleWritable message : messages){
                 msgCnt += 1;
             }
             if (vertex.getId().get() % 100000 == 0){
@@ -43,7 +43,7 @@ public class MessageBenchMark extends
             }
             MessageBenchMarkWorkerContext.messageReceived += msgCnt;
         }
-        LongWritable msg = new LongWritable(vertex.getId().get());
+        DoubleWritable msg = new DoubleWritable(vertex.getId().get());
         MessageBenchMarkWorkerContext.messageSent += vertex.getNumEdges();
         sendMessageToAllEdges(vertex, msg);
         //logger.info("Vertex [" + vertex.getId() + "] send to all edges " +  vertex.getId());
